# Optimize platformvm mempool peek operations (~20% less memory)