--- a/vms/platformvm/txs/mempool/mempool_benchmark_test.go
+++ b/vms/platformvm/txs/mempool/mempool_benchmark_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// Benchmarks marshal-ing "Version" message.
+// Benchmarks mempool PeekTxs with limits.
 //
 // e.g.,
 //