--- a/vms/platformvm/txs/mempool/mempool_test.go
+++ b/vms/platformvm/txs/mempool/mempool_test.go
@@ -14,9 +14,11 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/ava-labs/avalanchego/ids"
+	"github.com/ava-labs/avalanchego/utils/crypto/bls"
 	"github.com/ava-labs/avalanchego/utils/crypto/secp256k1"
 	"github.com/ava-labs/avalanchego/utils/timer/mockable"
 	"github.com/ava-labs/avalanchego/vms/components/avax"
+	"github.com/ava-labs/avalanchego/vms/platformvm/signer"
 	"github.com/ava-labs/avalanchego/vms/platformvm/txs"
 	"github.com/ava-labs/avalanchego/vms/secp256k1fx"
 )
@@ -172,6 +174,105 @@ func TestProposalTxsInMempool(t *testing.T) {
 	}
 }
 
+func TestPeekTxs(t *testing.T) {
+	require := require.New(t)
+
+	registerer := prometheus.NewRegistry()
+	mpool, err := NewMempool("mempool", registerer, &noopBlkTimer{})
+	require.NoError(err)
+
+	total := 100
+	decisionTxs, err := createTestDecisionTxs(total)
+	require.NoError(err)
+	stakerTxs, err := createTestAddPermissionlessValidatorTxs(total)
+	require.NoError(err)
+
+	// txs must not already there before we start
+	require.False(mpool.HasTxs())
+
+	oneDecisionTxSize, totalDecisionTxSize := 0, 0
+	for _, tx := range decisionTxs {
+		require.False(mpool.Has(tx.ID()))
+		require.NoError(mpool.Add(tx))
+
+		size := tx.Size()
+		if oneDecisionTxSize != 0 {
+			// assume all txs have the same size for the purpose of testing
+			require.Equal(oneDecisionTxSize, size)
+		} else {
+			oneDecisionTxSize = size
+		}
+
+		totalDecisionTxSize += oneDecisionTxSize
+	}
+
+	oneStakerTxSize, totalStakerTxSize := 0, 0
+	for _, tx := range stakerTxs {
+		require.False(mpool.Has(tx.ID()))
+		require.NoError(mpool.Add(tx))
+
+		size := tx.Size()
+		if oneStakerTxSize != 0 {
+			// assume all txs have the same size for the purpose of testing
+			require.Equal(oneStakerTxSize, size)
+		} else {
+			oneStakerTxSize = size
+		}
+
+		totalStakerTxSize += oneStakerTxSize
+	}
+
+	tt := []struct {
+		desc        string
+		maxTxBytes  int
+		expectedTxs int
+	}{
+		{
+			desc:        "peek txs with zero max tx bytes should return none",
+			maxTxBytes:  0,
+			expectedTxs: 0,
+		},
+		{
+			desc:        "peek txs with MaxInt should return all decision + staker txs",
+			maxTxBytes:  math.MaxInt,
+			expectedTxs: 2 * total,
+		},
+		{
+			desc:        "peek txs with totalDecisionTxSize + totalStakerTxSize should return all",
+			maxTxBytes:  totalDecisionTxSize + totalStakerTxSize,
+			expectedTxs: 2 * total,
+		},
+		{
+			desc:        "peek txs with totalDecisionTxSize should only return decision txs",
+			maxTxBytes:  totalDecisionTxSize,
+			expectedTxs: total,
+		},
+		{
+			desc:        "peek txs with totalDecisionTxSize - one decision tx size should return total - 1",
+			maxTxBytes:  totalDecisionTxSize - oneDecisionTxSize,
+			expectedTxs: total - 1,
+		},
+		{
+			desc:        "peek txs with half of totalDecisionTxSize should return total/2",
+			maxTxBytes:  totalDecisionTxSize / 2,
+			expectedTxs: total / 2,
+		},
+		{
+			desc:        "peek txs with totalDecisionTxSize + 3 staker tx size should return total + 3",
+			maxTxBytes:  totalDecisionTxSize + 3*oneStakerTxSize,
+			expectedTxs: total + 3,
+		},
+		{
+			desc:        "peek txs with totalDecisionTxSize + totalStakerTxSize/2 should return total + total/2",
+			maxTxBytes:  totalDecisionTxSize + totalStakerTxSize/2,
+			expectedTxs: total + total/2,
+		},
+	}
+	for i, tv := range tt {
+		require.Lenf(mpool.PeekTxs(tv.maxTxBytes), tv.expectedTxs, "[%d] %s", i, tv.desc)
+	}
+}
+
 func createTestDecisionTxs(count int) ([]*txs.Tx, error) {
 	decisionTxs := make([]*txs.Tx, 0, count)
 	for i := uint32(0); i < uint32(count); i++ {
@@ -240,4 +341,75 @@ func createTestProposalTxs(count int) ([]*txs.Tx, error) {
 		proposalTxs = append(proposalTxs, tx)
 	}
 	return proposalTxs, nil
+}
+
+func createTestAddPermissionlessValidatorTxs(count int) ([]*txs.Tx, error) {
+	var (
+		networkID = uint32(1337)
+		chainID   = ids.GenerateTestID()
+	)
+
+	// A BaseTx that passes syntactic verification.
+	validBaseTx := txs.BaseTx{
+		BaseTx: avax.BaseTx{
+			NetworkID:    networkID,
+			BlockchainID: chainID,
+		},
+	}
+
+	blsSK, err := bls.NewSecretKey()
+	if err != nil {
+		return nil, err
+	}
+	blsPOP := signer.NewProofOfPossession(blsSK)
+
+	signers := [][]*secp256k1.PrivateKey{{secp256k1.TestKeys()[0]}}
+
+	var clk mockable.Clock
+	tss := make([]*txs.Tx, 0, count)
+	for i := uint32(0); i < uint32(count); i++ {
+		utx := &txs.AddPermissionlessValidatorTx{
+			BaseTx: validBaseTx,
+			Validator: txs.Validator{
+				NodeID: ids.GenerateTestNodeID(),
+				Start:  uint64(clk.Time().Add(time.Duration(uint32(count)-i) * time.Second).Unix()),
+			},
+			Subnet: ids.GenerateTestID(),
+			Signer: blsPOP,
+			StakeOuts: []*avax.TransferableOutput{
+				{
+					Asset: avax.Asset{
+						ID: ids.GenerateTestID(),
+					},
+					Out: &secp256k1fx.TransferOutput{
+						Amt: 1,
+					},
+				},
+				{
+					Asset: avax.Asset{
+						ID: ids.GenerateTestID(),
+					},
+					Out: &secp256k1fx.TransferOutput{
+						Amt: 1,
+					},
+				},
+			},
+			ValidatorRewardsOwner: &secp256k1fx.OutputOwners{
+				Addrs:     []ids.ShortID{ids.GenerateTestShortID()},
+				Threshold: 1,
+			},
+			DelegatorRewardsOwner: &secp256k1fx.OutputOwners{
+				Addrs:     []ids.ShortID{ids.GenerateTestShortID()},
+				Threshold: 1,
+			},
+			DelegationShares: 20_000,
+		}
+
+		tx, err := txs.NewSigned(utx, txs.Codec, signers)
+		if err != nil {
+			return nil, err
+		}
+		tss = append(tss, tx)
+	}
+	return tss, nil
 }