## Why this should be merged

Simplify, optimize platform vm mempool peek operations.

## How this works

Previously, we just list and append all txs, and then iterate one by one with the max tx size limit. This PR introduces listing with limit at the heap level.

## How this was tested

Benchmarks + unit tests.

```bash
go install -v golang.org/x/tools/cmd/benchcmp@latest
go install -v golang.org/x/perf/cmd/benchstat@latest

go test -run=NONE -bench=BenchmarkPeekTxs > /tmp/cpu.before.txt
go test -run=NONE -bench=BenchmarkPeekTxs > /tmp/cpu.after.txt
benchcmp /tmp/cpu.before.txt /tmp/cpu.after.txt
benchstat -alpha 0.03 /tmp/cpu.before.txt /tmp/cpu.after.txt

go test -run=NONE -bench=BenchmarkPeekTxs -benchmem > /tmp/mem.before.txt
go test -run=NONE -bench=BenchmarkPeekTxs -benchmem > /tmp/mem.after.txt
benchcmp /tmp/mem.before.txt /tmp/mem.after.txt
benchstat -alpha 0.03 /tmp/mem.before.txt /tmp/mem.after.txt
```

```
benchmark               old ns/op     new ns/op     delta
BenchmarkPeekTxs-12     2540          1919          -24.45%

benchmark               old ns/op     new ns/op     delta
BenchmarkPeekTxs-12     2428          1796          -26.03%

benchmark               old allocs     new allocs     delta
BenchmarkPeekTxs-12     3              3              +0.00%

benchmark               old bytes     new bytes     delta
BenchmarkPeekTxs-12     13568         10240         -24.53%

pkg: github.com/ava-labs/avalanchego/vms/platformvm/txs/mempool
           │ /tmp/cpu.before.txt │       /tmp/cpu.after.txt        │
           │       sec/op        │    sec/op     vs base           │
PeekTxs-12          2.540µ ± ∞ ¹   1.919µ ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.03

goos: darwin
goarch: arm64
pkg: github.com/ava-labs/avalanchego/vms/platformvm/txs/mempool
           │ /tmp/mem.before.txt │       /tmp/mem.after.txt        │
           │       sec/op        │    sec/op     vs base           │
PeekTxs-12          2.428µ ± ∞ ¹   1.796µ ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.03

           │ /tmp/mem.before.txt │        /tmp/mem.after.txt        │
           │        B/op         │     B/op       vs base           │
PeekTxs-12         13.25Ki ± ∞ ¹   10.00Ki ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.03

           │ /tmp/mem.before.txt │       /tmp/mem.after.txt       │
           │      allocs/op      │  allocs/op   vs base           │
PeekTxs-12           3.000 ± ∞ ¹   3.000 ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
```