@gyuho 2023-05-02
For benchmarks, you can just reproduce with commenting out/in the following:

```go
func (m *mempool) PeekTxs(maxTxsBytes int) []*txs.Tx {
	// txs := m.unissuedDecisionTxs.List()
	// txs = append(txs, m.unissuedStakerTxs.List()...)

	// size := 0
	// for i, tx := range txs {
	// 	size += len(tx.Bytes())
	// 	if size > maxTxsBytes {
	// 		return txs[:i]
	// 	}
	// }
	// return txs

	txs, remaining := m.unissuedDecisionTxs.ListWithLimit(maxTxsBytes)
	if remaining <= 0 {
		return txs
	}
	txs2, _ := m.unissuedStakerTxs.ListWithLimit(remaining)
	return append(txs, txs2...)
}
```